From Yash/John meeting on 2020-12-02
We talked about Windows support, and how examples in docs have
backslashes which aren't compatable with windows \ cmd.exe. To
reduce the length of command line commands, and make it so we don't have
big differences in windows / linux commands in the docs, we will
implement @file.exension support for all arguments.

Load of files is mostly implemented, we first need to work on dict
support of @config classes. The first of those is Sources which needs to
be converted from a list to a regular config class. It will take a
config object, which will have a dict as a property.

The config code needs to be updated so that the dict property is where
stuff is put if it doesn't match any named property within the config
structure, this will be done by adding the field.fill parameter.

Signed-off-by: John Andersen <johnandersenpdx@gmail.com>