high level: Support for compressed files in _records_to_sources()
_records_to_sources() previously looked at the last suffix of the file
name to determine which source to load. This made it incompatible with
files that were compressed, i.e. .gz, .xz, etc.

This patch makes it so that we look at the first suffix. So if a
file name is some.csv.gz, the first suffix is csv and the second is gz.
We load the source using csv, the first suffix.

Signed-off-by: John Andersen <johnandersenpdx@gmail.com>