docs: tutorials: models: slr: Use features instead of feature
It was confusing to developers how they should switch from using a
single feature to multiple features. This patch updates the SLR model
tutorial to make the feature config property into features. Which is of
type Features rather than Feature.

The hope is that this will be a better place for developers to start
from because they are already using the features list.

Fixes: #903

Signed-off-by: John Andersen <johnandersenpdx@gmail.com>