df: base: Do not return None from auto definitioned functions
Returning None from an auto wrapped function is now treated the same as
if None was returned from a non-wrapped function. This way the input is
not sent to the input network.

We should probably add some boolean keyword argument to the op decorator
so that developers can toggle this behavior. They might mean to return
None.

Signed-off-by: John Andersen <johnandersenpdx@gmail.com>