util: config: inspect: Add make_config_inspect()
The make_config_inspect() function is the same as the
make_pytorch_config() function from the PyTorch models.

We move it to dffml.util.config.inspect:make_config_inspect() so that it
can be accessed from the main package.

The function can be used to create a config class out of any function
that has type hints on it's arguments, it also handles default
arguments.

Signed-off-by: John Andersen <johnandersenpdx@gmail.com>