df: memory: Instantiate Operation Implementations with their default config
Previously if no config was given in a dataflow for an operation it
would set the config for the OperationImplementation to BaseConfig. This
ignored the case when an OperationImplementation has a CONFIG property.

This patch changes the instantiation process so that Operation
Implementations now get instantiated with an instance of their config
object if they have one rather than an empty BaseConfig object if the
dataflow does not provide a config for the operation.

Signed-off-by: John Andersen <johnandersenpdx@gmail.com>